Going to be working on the trajectory arc tomorrow.
Ive put in a link in the raycasting script for me to have a look at tomorrow. I have also fixed an issue where the bullet would not shoot in the correct direction and die on impact but now it doesnt die on impact with player but with everything else. Will need to be changed to also not be hit by the npc.